refactor: Move `grpc/client.rs` to `grpc_client/sglang_scheduler.rs`
- Update references of grpc/client.rs
- Move logic of building proto-specific SamplingParams to sglang_scheduler.rs
