+++ sgl-router/src/grpc_client/mod.rs
@@ -1,0 +1,3 @@
+pub mod sglang_scheduler;
+
+pub use sglang_scheduler::{proto, SglangSchedulerClient};
